Reworked Vision Assist Features
Now there is an auto-aim and shoot while button held, and a driver aim assist (enabled by driver toggle) that only fires when the operator holds a separate button. Neither mode will fire unless the angle is within tolerance, but those have separate defined constants (so that the driver assist can shoot even if a bit mis aligned during defense). When in driver assist mode, the shooter is set to 2000rpm until the mode is disabled/the shooting button is pressed.